Split theme appearance helpers into focused files
The Swift file length budget guard fails on the committed appearance work because GhosttyTerminalView and WorkspaceContentView absorb both the runtime notification dispatcher and the resolved-appearance helpers, while GhosttyConfigTests inherits the matching coverage.

This extracts the Ghostty default-background notification types into Sources/GhosttyTerminalAppearance.swift, moves resolveGhosttyAppearanceConfig and ghosttyAppearanceSignature into Sources/WorkspaceAppearanceResolution.swift, and splits the corresponding tests into cmuxTests/GhosttyNotificationDispatcherTests.swift and cmuxTests/WorkspaceAppearanceConfigResolutionTests.swift. The Xcode project picks up the new files and the budget tsv now reflects the post-split line counts.

Constraint: Move only what unblocks the budget guard; do not touch behavior or rename public surface.

Rejected: Bump the budget for the committed deltas without splitting | postpones the work and grows the tracked debt.

Confidence: high

Scope-risk: narrow

Directive: Keep appearance dispatcher and workspace appearance resolution in their own files so future growth lands away from the chrome megafiles.

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Not-tested: Local XCTest per repository policy; CI will re-verify.
