+++ cmuxTests/GhosttyConfigTests.swift
@@ -874,73 +874,6 @@ final class WindowChromeSeparatorColorTests: XCTestCase {
     }
 }
 
-final class WorkspaceAppearanceConfigResolutionTests: XCTestCase {
-    func testResolvedAppearanceConfigPrefersGhosttyRuntimeAppearanceOverLoadedConfig() {
-        guard let loadedBackground = NSColor(hex: "#112233"),
-              let runtimeBackground = NSColor(hex: "#FDF6E3"),
-              let loadedForeground = NSColor(hex: "#EEEEEE"),
-              let runtimeForeground = NSColor(hex: "#4A4543"),
-              let loadedCursor = NSColor(hex: "#DDDDDD"),
-              let runtimeCursor = NSColor(hex: "#3A3432"),
-              let loadedCursorText = NSColor(hex: "#111111"),
-              let runtimeCursorText = NSColor(hex: "#F7F7F7"),
-              let loadedSelectionBackground = NSColor(hex: "#222222"),
-              let runtimeSelectionBackground = NSColor(hex: "#A5A2A2"),
-              let loadedSelectionForeground = NSColor(hex: "#EEEEEE"),
-              let runtimeSelectionForeground = NSColor(hex: "#4A4543") else {
-            XCTFail("Expected valid test colors")
-            return
-        }
-
-        var loaded = GhosttyConfig()
-        loaded.backgroundColor = loadedBackground
-        loaded.foregroundColor = loadedForeground
-        loaded.cursorColor = loadedCursor
-        loaded.cursorTextColor = loadedCursorText
-        loaded.selectionBackground = loadedSelectionBackground
-        loaded.selectionForeground = loadedSelectionForeground
-        loaded.unfocusedSplitOpacity = 0.42
-
-        let resolved = WorkspaceContentView.resolveGhosttyAppearanceConfig(
-            loadConfig: { loaded },
-            defaultBackground: { runtimeBackground },
-            defaultForeground: { runtimeForeground },
-            defaultCursor: { runtimeCursor },
-            defaultCursorText: { runtimeCursorText },
-            defaultSelectionBackground: { runtimeSelectionBackground },
-            defaultSelectionForeground: { runtimeSelectionForeground }
-        )
-
-        XCTAssertEqual(resolved.backgroundColor.hexString(), "#FDF6E3")
-        XCTAssertEqual(resolved.foregroundColor.hexString(), "#4A4543")
-        XCTAssertEqual(resolved.cursorColor.hexString(), "#3A3432")
-        XCTAssertEqual(resolved.cursorTextColor.hexString(), "#F7F7F7")
-        XCTAssertEqual(resolved.selectionBackground.hexString(), "#A5A2A2")
-        XCTAssertEqual(resolved.selectionForeground.hexString(), "#4A4543")
-        XCTAssertEqual(resolved.unfocusedSplitOpacity, 0.42, accuracy: 0.0001)
-    }
-
-    func testResolvedAppearanceConfigPrefersExplicitBackgroundOverride() {
-        guard let loadedBackground = NSColor(hex: "#112233"),
-              let runtimeBackground = NSColor(hex: "#FDF6E3"),
-              let explicitOverride = NSColor(hex: "#272822") else {
-            XCTFail("Expected valid test colors")
-            return
-        }
-
-        var loaded = GhosttyConfig()
-        loaded.backgroundColor = loadedBackground
-
-        let resolved = WorkspaceContentView.resolveGhosttyAppearanceConfig(
-            backgroundOverride: explicitOverride,
-            loadConfig: { loaded },
-            defaultBackground: { runtimeBackground }
-        )
-
-        XCTAssertEqual(resolved.backgroundColor.hexString(), "#272822")
-    }
-}
-
 @MainActor
 final class WorkspaceChromeColorTests: XCTestCase {
     func testBonsplitChromeHexIncludesAlphaWhenTranslucent() {
@@ -1976,95 +1909,6 @@ final class NotificationBurstCoalescerTests: XCTestCase {
 
         wait(for: [expectation], timeout: 1.0)
         XCTAssertEqual(flushCount, 2)
-    }
-}
-
-final class GhosttyDefaultBackgroundNotificationDispatcherTests: XCTestCase {
-    func testSignalCoalescesBurstToLatestBackground() {
-        guard let dark = NSColor(hex: "#272822"),
-              let light = NSColor(hex: "#FDF6E3") else {
-            XCTFail("Expected valid test colors")
-            return
-        }
-
-        let expectation = expectation(description: "coalesced notification")
-        expectation.expectedFulfillmentCount = 1
-        var postedUserInfos: [[AnyHashable: Any]] = []
-
-        let dispatcher = GhosttyDefaultBackgroundNotificationDispatcher(
-            delay: 0.01,
-            postNotification: { userInfo in
-                postedUserInfos.append(userInfo)
-                expectation.fulfill()
-            }
-        )
-
-        DispatchQueue.main.async {
-            dispatcher.signal(backgroundColor: dark, opacity: 0.95, eventId: 1, source: "test.dark")
-            dispatcher.signal(backgroundColor: light, opacity: 0.75, eventId: 2, source: "test.light")
-        }
-
-        wait(for: [expectation], timeout: 1.0)
-        XCTAssertEqual(postedUserInfos.count, 1)
-        XCTAssertEqual(
-            (postedUserInfos[0][GhosttyNotificationKey.backgroundColor] as? NSColor)?.hexString(),
-            "#FDF6E3"
-        )
-        XCTAssertEqual(
-            postedOpacity(from: postedUserInfos[0][GhosttyNotificationKey.backgroundOpacity]),
-            0.75,
-            accuracy: 0.0001
-        )
-        XCTAssertEqual(
-            (postedUserInfos[0][GhosttyNotificationKey.backgroundEventId] as? NSNumber)?.uint64Value,
-            2
-        )
-        XCTAssertEqual(
-            postedUserInfos[0][GhosttyNotificationKey.backgroundSource] as? String,
-            "test.light"
-        )
-    }
-
-    func testSignalAcrossSeparateBurstsPostsMultipleNotifications() {
-        guard let dark = NSColor(hex: "#272822"),
-              let light = NSColor(hex: "#FDF6E3") else {
-            XCTFail("Expected valid test colors")
-            return
-        }
-
-        let expectation = expectation(description: "two notifications")
-        expectation.expectedFulfillmentCount = 2
-        var postedHexes: [String] = []
-
-        let dispatcher = GhosttyDefaultBackgroundNotificationDispatcher(
-            delay: 0.01,
-            postNotification: { userInfo in
-                let hex = (userInfo[GhosttyNotificationKey.backgroundColor] as? NSColor)?.hexString() ?? "nil"
-                postedHexes.append(hex)
-                expectation.fulfill()
-            }
-        )
-
-        DispatchQueue.main.async {
-            dispatcher.signal(backgroundColor: dark, opacity: 1.0, eventId: 1, source: "test.dark")
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.05) {
-                dispatcher.signal(backgroundColor: light, opacity: 1.0, eventId: 2, source: "test.light")
-            }
-        }
-
-        wait(for: [expectation], timeout: 1.0)
-        XCTAssertEqual(postedHexes, ["#272822", "#FDF6E3"])
-    }
-
-    private func postedOpacity(from value: Any?) -> Double {
-        if let value = value as? Double {
-            return value
-        }
-        if let value = value as? NSNumber {
-            return value.doubleValue
-        }
-        XCTFail("Expected background opacity payload")
-        return -1
     }
 }
